Mostly fixed collision detection. Implemented a collision-detection heuristic to minimise computation. Added two more level files. Added ladders. Added textures for previously unused tiles (sewer walls, building walls, platforms, ladders Added Map class stub that will represent the overworld map. Added staticVerticalMotion function to Physics engine for use with ladders (and later cannons). Added collision detection for ladders
TODO:
Fix ladders so that they function as platform tiles when the W button is released
	This should also allow ladders to be shortened to platform height, rather than 2 blocks above currently
Fix collision detection errors related to bottom-of-tile collision (currently the character is moved right one tile when colliding with a tile from below)
Implement cannons!
Add the next 3 levels
Implement the overworld map
Reset levelx to 0 as a default (rather than the 2 used for debugging)
Create all necessary textures.
Create the backgrounds and split them into layers and segments
Change the textures list to a dictionary for text-based lookup
	Current system does not allow different positions for textures when they are initially loaded)

Signed-off-by: Andrew Pigram <amp942@uowmail.edu.au>